Use Long as migration version (#4297)
* Use long as migration version

* Add integration test

* Bump to sql-psi 0.4.5

* Fix int usage

* Fix dialect test

* Fix dialect tests

* Fix schema version

* Fix schema version

* Fix tests

---------

+++ drivers/sqljs-driver/src/jsTest/kotlin/com/squareup/sqldelight/drivers/sqljs/JsQueryTest.kt
@@ -27,7 +27,7 @@ class JsQueryTest {
   }
 
   private val schema = object : SqlSchema<QueryResult.Value<Unit>> {
-    override val version: Int = 1
+    override val version: Long = 1
 
     override fun create(driver: SqlDriver): QueryResult.Value<Unit> {
       driver.execute(
@@ -45,8 +45,8 @@ class JsQueryTest {
 
     override fun migrate(
       driver: SqlDriver,
-      oldVersion: Int,
-      newVersion: Int,
+      oldVersion: Long,
+      newVersion: Long,
       vararg callbacks: AfterVersion,
     ): QueryResult.Value<Unit> {
       // No-op.
